feat: separate Task Manager into `Task` module
